move did-define-property mixin into DS.Model directly
